+++ services/api/src/modules/integrations/index.ts
@@ -1,0 +1,4 @@
+export * from './integrations.module';
+export * from './integrations.service';
+export * from './interfaces';
+export * from './providers';