+++ services/api/src/modules/integrations/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './sms-provider.interface';
+export * from './notification-provider.interface';
+export * from './accounting-adapter.interface';