+++ services/api/src/modules/integrations/providers/index.ts
@@ -1,0 +1,3 @@
+export * from './noop-sms.provider';
+export * from './noop-notification.provider';
+export * from './noop-accounting.adapter';